Update solo-staker-migration.md
Added a short eth-docker option to section 4 as the software is widely used and not mentioned in here. Their implementation is quite handy and I took hours to figure this out. Having this in the original guide is a must ;) 

Additionally, changed a wrong link in section 5 about promoting the minipool.